Refactored the code
Extracted the subgraph comparison algorithm from the GUI module.
Made some code a bit nicer:
  - the thread classes got getters which join themselves.
  - there is no static reference to the RootLayoutController class anymore.
Added some comments.
