+++ PL2/PL2-gui/src/main/java/nl/tudelft/pl2016gr2/gui/view/selection/SelectionManager.java
@@ -10,6 +10,9 @@ import javafx.scene.Node;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.Region;
 import javafx.util.Duration;
+import nl.tudelft.pl2016gr2.gui.view.RootLayoutController;
+
+import java.util.ArrayList;
 
 /**
  * This class manages the currently selected node. It makes sure the correct data is displayed and
@@ -19,6 +22,7 @@ import javafx.util.Duration;
  */
 public class SelectionManager {
 
+  private final RootLayoutController rootLayoutController;
   private final Pane selectionDescriptionPane;
   private final Region background;
   private DescriptionPane contentPane;
@@ -28,10 +32,13 @@ public class SelectionManager {
   /**
    * Create a selection manager.
    *
+   * @param rootLayoutController     the root layout controller class.
    * @param selectionDescriptionPane the pane in which to draw information about selected items.
    * @param background               the background pane which is positioned behind the description.
    */
-  public SelectionManager(Pane selectionDescriptionPane, Region background) {
+  public SelectionManager(RootLayoutController rootLayoutController, Pane selectionDescriptionPane,
+      Region background) {
+    this.rootLayoutController = rootLayoutController;
     this.selectionDescriptionPane = selectionDescriptionPane;
     this.background = background;
 
@@ -70,6 +77,10 @@ public class SelectionManager {
     }
   }
 
+  protected void drawGraph(ArrayList<String> topGenomes, ArrayList<String> bottomGenomes) {
+    rootLayoutController.drawGraph(topGenomes, bottomGenomes);
+  }
+
   /**
    * Set the content of the selection description pane.
    *
@@ -77,7 +88,7 @@ public class SelectionManager {
    */
   private void createDescription(ISelectable selected) {
     createNewContentPane();
-    Node description = selected.getSelectionInfo().getNode();
+    Node description = selected.getSelectionInfo(this).getNode();
     contentPane.getChildren().add(description);
     contentPane.setOpacity(0);
     timeline = new Timeline();
